dropping ast-precompiler
