+++ packages/core/src/template-compiler.ts
@@ -215,6 +215,26 @@ export default class TemplateCompiler {
   baseDir() {
     return join(__dirname, '..');
   }
+
+  // tests for the classic ember-cli-htmlbars-inline-precompile babel plugin
+  static isInlinePrecompilePlugin(item: PluginItem) {
+    if (typeof item === 'string') {
+      return matchesSourceFile(item);
+    }
+    if (hasProperties(item) && (item as any)._parallelBabel) {
+      return matchesSourceFile((item as any)._parallelBabel.requireFile);
+    }
+    if (Array.isArray(item) && item.length > 0) {
+      if (typeof item[0] === 'string') {
+        return matchesSourceFile(item[0]);
+      }
+      if (hasProperties(item[0]) && (item[0] as any)._parallelBabel) {
+        return matchesSourceFile((item[0] as any)._parallelBabel.requireFile);
+      }
+    }
+    return false;
+  }
+
 }
 
 class TemplateCompileTree extends Filter {
@@ -243,3 +263,12 @@ class TemplateCompileTree extends Filter {
     return join(__dirname, '..');
   }
 }
+
+function matchesSourceFile(filename: string) {
+  return /babel-plugin-htmlbars-inline-precompile\/(index|lib\/require-from-worker)\.js$/.test(filename);
+}
+
+function hasProperties(item: any) {
+  return item && (typeof item === 'object' || typeof item === 'function');
+}
+
